+++ src/MagicOnion.GeneratorCore/Generator/CodeTemplate.tt
@@ -3,7 +3,7 @@
 <#@ import namespace="System.Linq" #>
 <#@ import namespace="System.Text" #>
 <#@ import namespace="System.Collections.Generic" #>
-<#@ import namespace="MagicOnion.CodeAnalysis" #>
+<#@ import namespace="MagicOnion.Generator.CodeAnalysis" #>
 #pragma warning disable 618
 #pragma warning disable 612
 #pragma warning disable 414
