+++ src/MagicOnion.GeneratorCore/Utils/RoslynExtensions.cs
@@ -1,10 +1,10 @@
+using System.Collections.Generic;
+using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.Collections.Generic;
-using System.Linq;
 
-namespace MagicOnion
+namespace MagicOnion.Generator.Utils
 {
     // Utility and Extension methods for Roslyn
     internal static class RoslynExtensions
